Trait pallet_relayer_registry::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Currency: ReservableCurrency<Self::AccountId>;
    type BasicDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type FieldDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type MaxAdditionalFields: Get<u32>;
    type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type Currency: ReservableCurrency<Self::AccountId>

The currency trait.

type BasicDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The amount held on deposit for a registered identity

type FieldDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The amount held on deposit per additional field for a registered identity.

type MaxAdditionalFields: Get<u32>

Maximum number of additional fields that may be stored in an ID. Needed to bound the I/O required to access an identity, but can be pretty high.

type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin which may forcibly set or remove a name. Root can always do this.

type WeightInfo: WeightInfo